refactor: 캐러셀 폴더 생성
