+++ src/pages/List.jsx
@@ -1,4 +1,4 @@
-import Carousel from "../components/Carousel";
+import Carousel from "../components/Carousel/Carousel";
 
 function List() {
   return (
